+++ util/clixon_util_ssl.c
@@ -539,7 +539,7 @@ main(int    argc,
     }
     if (clixon_event_reg_fd(ss, ssl_input_cb, sd, "ssl socket") < 0)
 	goto done;
-    if (clixon_event_loop() < 0)
+    if (clixon_event_loop(h) < 0)
 	goto done;
     retval = 0;
  done: